==> 02-data-preparation.ipynb <==
## **2. Data Preparation**

#### **Step 2.1 Read Songs**

import os
import pandas as pd
from slugify import slugify

artist_name = 'Avenged Sevenfold'
artist_slug = slugify(artist_name)

df = pd.read_csv(f'unduhan/{artist_slug}.csv')
df = df.sort_values('Judul')

df.head()

#### **Step 2.2 Merge Lyrics**

folder_path = f'unduhan/{artist_slug}'
lyrics = []

for file_name in os.listdir(folder_path):
    if file_name.endswith('.txt'):
        file_path = os.path.join(folder_path, file_name)
        with open(file_path, 'r') as file:
            text = file.read()
            lyrics.append(text)

df['Lirik'] = lyrics
df = df.drop(columns=['year', 'type', 'url'])
df = df.dropna()

df.tail()

Simpan sebagai file baru dengan separator `";"` untuk menghindari penggunaan koma `(,)` dalam lirik.

<small>Tentunya kami sudah memastikan tidak ada penggunaan `";"` pada lirik yang diunduh</small>

df.to_csv(f'unduhan/{artist_slug}-with-lyrics.csv', sep=";")

#### **(Opsional)**

Kami berencana untuk melakukan analisis kata setiap masing-masing albumnya.

from tabulate import tabulate

album_list = df['album'].unique()

for album in album_list:
    album_data = df[df['album'] == album]
    album_data.to_csv(f'unduhan/{artist_slug}/album/{album}.csv', index=False, sep=";")
    
table = []
for i, j in enumerate(album_list, start=1):
    table.append([i, j])

headers = ["No", "Nama Album"]
print(tabulate(table, headers, tablefmt="grid"))
