==> 04-preprocessing.ipynb <==
## **4. Preprocessing**

#### **Step 1: File Lirik (CSV)**

import pandas as pd
import pickle
from slugify import slugify

artist_name = 'Avenged Sevenfold'
artist_slug = slugify(artist_name)

lyric_corpus = pd.read_csv(f'unduhan/{artist_slug}-with-lyrics.csv', sep=";")
lyric_corpus = lyric_corpus.sort_values('Judul')
lyric_corpus = lyric_corpus.drop(columns=['Unnamed: 0', 'Judul', 'album'])

lyric_corpus.head()

#### **Step 2: Tokensasi Lirik**

from nltk.tokenize import RegexpTokenizer

lyric_corpus_tokenized = []
tokenizer = RegexpTokenizer(r'\w+')
for lyric in lyric_corpus['Lirik']:
    tokenized_lyric = tokenizer.tokenize(lyric.lower())
    lyric_corpus_tokenized.append(tokenized_lyric)
    
lyric_corpus['Token'] = lyric_corpus_tokenized
lyric_corpus.tail()

#### **Step 3: Hapus Angka dan Kata < 3 karakter**

Kami kemudian menghapus token numerik serta token yang berisi kurang dari 3 karakter (seperti bunyi melodi 'oh', 'na', 'la', 'da') yang cukup bisa mendistorsi hasil dari topik modelling.

for s, song in enumerate(lyric_corpus_tokenized):
    filtered_song = []    
    for token in song:
        if len(token) > 2 and not token.isnumeric():
            filtered_song.append(token)
    lyric_corpus_tokenized[s] = filtered_song

#### **Step 4: Lematisasi**

from nltk.stem.wordnet import WordNetLemmatizer

lemmatizer = WordNetLemmatizer()
for s,song in enumerate(lyric_corpus_tokenized):
    lemmatized_tokens = []
    for token in song:
        lemmatized_tokens.append(lemmatizer.lemmatize(token))
    lyric_corpus_tokenized[s] = lemmatized_tokens

#### **Step 5: Hapus Stop Words**

Terakhir, semua kata yang tidak memiliki arti khusus dalam Topic Modeling akan kami hapus untuk mengurangi noise. Kami menggunakan stopwords dari NLTK dalam bahasa inggris mengingat seluruh lirik lagu yang ada menggunakan Bahasa Inggris.

from nltk.corpus import stopwords

stop_words = stopwords.words('english')
new_stop_words = ['ooh','yeah','hey','whoa','woah', 'ohh', 'hmm']
stop_words.extend(new_stop_words)
for s,song in enumerate(lyric_corpus_tokenized):
    filtered_text = []    
    for token in song:
        if token not in stop_words:
            filtered_text.append(token)
    lyric_corpus_tokenized[s] = filtered_text

#### **Step 6: Pickle It!**
Kami melakukan proses bundling token kedalam 1 file pickle untuk digunakan kemudian.

with open('pickle/tokenized.pkl', 'wb') as f:
    pickle.dump(lyric_corpus_tokenized, f)

Sampai pada proses ini, output yang kami dapat adalah file pickle `tokenized.pkl`.

print(lyric_corpus_tokenized)
