==> 06-evaluate.ipynb <==
## **6. Evaluasi**

Dalam *Topic Modelling*, untuk menilai kualitas suatu model umumnya dapat menggunakan skor koherensi. Dimana, koherensi topik ini akan mengukur sejauh mana topik-topik yang dihasilkan konsisten dan bermakna secara semantik.

Dalam implementasinya, kami menggunakan pustaka **gensim** dengan **C_v** sebagai metodenya.

import pickle

with open('pickle/lda_model.pkl', 'rb') as f:
    lda_model = pickle.load(f)

with open('pickle/tokenized.pkl', 'rb') as f:
    texts = pickle.load(f)

with open('pickle/dictionary.pkl', 'rb') as f:
    dictionary = pickle.load(f)

from gensim.models.coherencemodel import CoherenceModel

coherencemodel = CoherenceModel(
    model=lda_model, 
    texts=texts, 
    dictionary=dictionary, 
    coherence='c_v')

coherence = coherencemodel.get_coherence()

print(f"Coherence Score: {coherence}")

Dan menghasilkan nilai Coherence Score sebesar **0.75**.
